Create Distribution count sort.cpp
Distribution count sort program implemented in cpp, whose efficiency is o(n), which is linear. Hence this algo is more efficient that all other algos till now what you have studied, but constraint for  this is that we have use the values in such a way that all the values must be linearly distributed. For random numbers this algo fails.